+++ trixnity-client/src/commonTest/kotlin/net/folivo/trixnity/client/store/RoomUserStoreTest.kt
@@ -6,6 +6,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.first
+import kotlinx.datetime.Clock
 import net.folivo.trixnity.client.MatrixClientConfiguration
 import net.folivo.trixnity.client.flatten
 import net.folivo.trixnity.client.mocks.RepositoryTransactionManagerMock
@@ -35,7 +36,8 @@ class RoomUserStoreTest : ShouldSpec({
             RepositoryTransactionManagerMock(),
             MatrixClientConfiguration(),
             ObservableCacheStatisticCollector(),
-            storeScope
+            storeScope,
+            Clock.System,
         )
     }
     afterTest {